move apply_tools from proscenium.complete to example.tools.util
