+++ example/tools/answer_question.py
@@ -8,8 +8,9 @@ from gofannon.basic_math.multiplication import Multiplication
 from gofannon.basic_math.division import Division
 
 from proscenium.display import header
-from proscenium.complete import apply_tools
 from proscenium.invoke import process_tools
+
+import example.tools.util as util
 
 tools = [Addition, Subtraction, Multiplication, Division]
 
@@ -24,9 +25,9 @@ question = Prompt.ask(
     default = "What is 33312-457? And what is 3+3?"
     )
 
-answer = apply_tools(
+answer = util.apply_tools(
     model_id = "ollama:llama3.2",
-    model_id = "ollama:granite3.2",
+    # model_id = "ollama:granite3.2",
     # model_id = "openai:gpt-4o",
     system_message = "Perform any referenced arithmetic.",
     message = question,
