+++ example/tools/util.py
@@ -1,12 +1,14 @@
 
 from typing import List
 
-from aisuite import Client
+from rich import print
+from rich.panel import Panel
+from rich.text import Text
 from thespian.actors import Actor
 
 from gofannon.base import BaseTool
 
-from proscenium.complete import apply_tools
+from proscenium.complete import complete_for_tool_applications, evaluate_tool_calls, complete_with_tool_results
 from proscenium.invoke import process_tools
 
 def tool_applier_actor_class(
@@ -34,3 +36,57 @@ def tool_applier_actor_class(
             self.send(sender, response)
 
     return ToolApplier
+
+def apply_tools(
+    model_id: str,
+    system_message: str,
+    message: str,
+    tool_desc_list: list,
+    tool_map: dict,
+    temperature: float = 0.75,
+    rich_output: bool = False
+    ) -> str:
+
+    messages = [
+        {"role": "system", "content": system_message},
+        {"role": "user", "content": message},
+    ]
+
+    response = complete_for_tool_applications(
+        model_id,
+        messages,
+        tool_desc_list,
+        temperature,
+        rich_output
+    )
+
+    tool_call_message = response.choices[0].message
+
+    if tool_call_message.tool_calls is None:
+
+        if rich_output:
+            print(Panel(Text(str(tool_call_message.content)), title="Tool Application Response"))
+
+        print("No tool applications detected")
+
+        return tool_call_message.content
+
+    else:
+
+        if rich_output:
+            print(Panel(Text(str(tool_call_message)), title="Tool Application Response"))
+
+        tool_evaluation_messages = evaluate_tool_calls(
+            tool_call_message, tool_map, rich_output)
+
+        result = complete_with_tool_results(
+            model_id,
+            messages,
+            tool_call_message,
+            tool_evaluation_messages,
+            tool_desc_list,
+            temperature,
+            rich_output
+        )
+
+        return result
