+++ proscenium/complete.py
@@ -200,58 +200,4 @@ def complete_with_tool_results(
 
     return response.choices[0].message.content
 
-def apply_tools(
-    model_id: str,
-    system_message: str,
-    message: str,
-    tool_desc_list: list,
-    tool_map: dict,
-    temperature: float = 0.75,
-    rich_output: bool = False
-    ) -> str:
-
-    messages = [
-        {"role": "system", "content": system_message},
-        {"role": "user", "content": message},
-    ]
-
-    response = complete_for_tool_applications(
-        model_id,
-        messages,
-        tool_desc_list,
-        temperature,
-        rich_output
-    )
-
-    tool_call_message = response.choices[0].message
-
-    if tool_call_message.tool_calls is None:
-
-        if rich_output:
-            print(Panel(Text(str(tool_call_message.content)), title="Tool Application Response"))
-
-        print("No tool applications detected")
-
-        return tool_call_message.content
-
-    else:
-
-        if rich_output:
-            print(Panel(Text(str(tool_call_message)), title="Tool Application Response"))
-
-        tool_evaluation_messages = evaluate_tool_calls(
-            tool_call_message, tool_map, rich_output)
-
-        result = complete_with_tool_results(
-            model_id,
-            messages,
-            tool_call_message,
-            tool_evaluation_messages,
-            tool_desc_list,
-            temperature,
-            rich_output
-        )
-
-        return result
-
